Fix screen ECG
