+++ app/src/data_module.c
@@ -312,8 +312,8 @@ void data_thread(void)
 
             if (settings_send_ble_enabled)
             {
-                ble_ecg_notify(ecg_bioz_sensor_sample.ecg_samples, ecg_bioz_sensor_sample.ecg_num_samples);
-                ble_bioz_notify(ecg_bioz_sensor_sample.bioz_sample, ecg_bioz_sensor_sample.bioz_num_samples);
+                //ble_ecg_notify(ecg_bioz_sensor_sample.ecg_samples, ecg_bioz_sensor_sample.ecg_num_samples);
+                //ble_bioz_notify(ecg_bioz_sensor_sample.bioz_sample, ecg_bioz_sensor_sample.bioz_num_samples);
                 // b_notify(ecg_bioz_sensor_sample.bioz_sample);
 
                 /*resp_sample_buffer[resp_sample_buffer_count++] = ecg_bioz_sensor_sample.bioz_sample;
@@ -328,7 +328,7 @@ void data_thread(void)
             /***** Send to USB if enabled *****/
             if (settings_send_usb_enabled)
             {
-                if (settings_data_format == DATA_FMT_OPENVIEW)
+                /*if (settings_data_format == DATA_FMT_OPENVIEW)
                 {
                     sendData(sensor_sample.ecg_sample, sensor_sample.bioz_sample, sensor_sample.raw_red, sensor_sample.raw_ir,
                              (double)(sensor_sample.temp / 10.00), computed_data.hr, computed_data.rr, computed_data.spo2, sensor_sample._bioZSkipSample);
@@ -341,7 +341,7 @@ void data_thread(void)
                 {
                     send_data_ov3_format(ecg_bioz_sensor_sample.ecg_samples, ecg_bioz_sensor_sample.bioz_sample, ecg_bioz_sensor_sample.ecg_samples,
                                          ecg_bioz_sensor_sample.ecg_samples, sensor_sample.temp, computed_data.hr, computed_data.rr, computed_data.spo2, sensor_sample._bioZSkipSample);
-                }
+                }*/
             }
 
 #ifdef CONFIG_HEALTHYPI_DISPLAY_ENABLED
