+++ app/src/ui/screens/scr_ecg.c
@@ -3,7 +3,6 @@
 #include <zephyr/device.h>
 #include <lvgl.h>
 #include <stdio.h>
-#include <zephyr/smf.h>
 #include <app_version.h>
 
 #include "sys_sm_module.h"
@@ -16,113 +15,132 @@
 static lv_obj_t *scr_ecg;
 
 // GUI Charts
-static lv_obj_t *chart1;
-static lv_chart_series_t *ser1;
+static lv_obj_t *chart_ecg;
+static lv_chart_series_t *ser_ecg;
 
-static bool chart1_update = true;
+static bool chart_ecg_update = true;
 
-static float y1_max = 0;
-static float y1_min = 10000;
-
+//Chart Variables
+static float y_max_ecg = 0;
+static float y_min_ecg = 10000;
 static float gx = 0;
 
 // GUI Styles
 extern lv_style_t style_hr;
+extern lv_style_t style_spo2;
+extern lv_style_t style_rr;
+extern lv_style_t style_temp;
 extern lv_style_t style_sub;
 
 extern int curr_screen;
 
+#define DISP_WINDOW_SIZE_ECG 390
+
 void draw_scr_ecg(enum scroll_dir m_scroll_dir)
 {
-
     scr_ecg = lv_obj_create(NULL);
-    draw_scr_ecg_footer(scr_ecg);
+    draw_scr_home_footer(scr_ecg);
     draw_header(scr_ecg, true);
 
-    // lv_obj_add_style(scr_ecg, &style_welcome_scr_bg, 0);
+    chart_ecg = lv_chart_create(scr_ecg);
+    lv_obj_set_pos(chart_ecg, 10, 25);
 
-    // lv_group_t *g1 = lv_group_create();
+    lv_obj_set_size(chart_ecg, 400, 100);
+    lv_obj_set_style_bg_color(chart_ecg, lv_color_black(), LV_STATE_DEFAULT);
 
-    // Create Chart 1
-    chart1 = lv_chart_create(scr_ecg);
-    lv_obj_set_size(chart1, 460, 180);
-    lv_obj_set_style_bg_color(chart1, lv_color_black(), LV_STATE_DEFAULT);
+    lv_obj_set_style_size(chart_ecg, 0, LV_PART_INDICATOR);
+    lv_chart_set_point_count(chart_ecg, DISP_WINDOW_SIZE);
+    lv_chart_set_range(chart_ecg, LV_CHART_AXIS_PRIMARY_Y, -1000, 1000);
+    lv_chart_set_div_line_count(chart_ecg, 0, 0);
+    lv_chart_set_update_mode(chart_ecg, LV_CHART_UPDATE_MODE_CIRCULAR);
 
-    lv_obj_set_style_size(chart1, 0, LV_PART_INDICATOR);
-    lv_chart_set_point_count(chart1, DISP_WINDOW_SIZE);
-    lv_chart_set_range(chart1, LV_CHART_AXIS_PRIMARY_Y, -1000, 1000);
-    lv_chart_set_div_line_count(chart1, 0, 0);
-    lv_chart_set_update_mode(chart1, LV_CHART_UPDATE_MODE_CIRCULAR);
+    ser_ecg = lv_chart_add_series(chart_ecg, lv_palette_main(LV_PALETTE_RED), LV_CHART_AXIS_PRIMARY_Y);
 
-    lv_obj_set_pos(chart1, 10, 25);
+    //lv_obj_t *label_chart_title = lv_label_create(scr_ecg);
 
-    lv_obj_t *label_chart_title = lv_label_create(scr_ecg);
-
-    lv_obj_align_to(label_chart_title, chart1, LV_ALIGN_OUT_BOTTOM_MID, 0, 0);
-    lv_label_set_text(label_chart_title, "Showing ECG");
+    //lv_obj_align_to(label_chart_title, chart_ecg, LV_ALIGN_OUT_BOTTOM_MID, 0, 0);
+    //lv_label_set_text(label_chart_title, "Showing ECG");
 
     /*}
     else if (m_data_type == HPI_SENSOR_DATA_PPG)
     {
-        ser1 = lv_chart_add_series(chart1, lv_palette_main(LV_PALETTE_YELLOW), LV_CHART_AXIS_PRIMARY_Y);
+        ser_ecg = lv_chart_add_series(chart_ecg, lv_palette_main(LV_PALETTE_YELLOW), LV_CHART_AXIS_PRIMARY_Y);
         curr_screen = HPI_DISP_SCR_PPG;
         lv_label_set_text(label_chart_title, "Showing PPG");
     }
     else if (m_data_type == HPI_SENSOR_DATA_RESP)
     {
-        ser1 = lv_chart_add_series(chart1, lv_palette_main(LV_PALETTE_BLUE), LV_CHART_AXIS_PRIMARY_Y);
+        ser_ecg = lv_chart_add_series(chart_ecg, lv_palette_main(LV_PALETTE_BLUE), LV_CHART_AXIS_PRIMARY_Y);
         curr_screen = HPI_DISP_SCR_RESP;
         lv_label_set_text(label_chart_title, "Showing RESP");
     }*/
-    ser1 = lv_chart_add_series(chart1, lv_palette_main(LV_PALETTE_RED), LV_CHART_AXIS_PRIMARY_Y);
-
-    curr_screen = SCR_ECG;
+   
+    curr_screen = SCR_HOME;
 
     hpi_show_screen(scr_ecg, m_scroll_dir);
 }
 
-static void hpi_ecg_add_samples(int num_samples)
+static void hpi_ecg_disp_add_samples(int num_samples)
 {
     gx += num_samples;
 }
 
-static void hpi_ecg_do_set_scale()
+static void hpi_ecg_disp_do_set_scale(int disp_window_size)
 {
-    if (gx >= DISP_WINDOW_SIZE)
+    if (gx >= (disp_window_size))
     {
-        if (chart1_update == true)
-            lv_chart_set_range(chart1, LV_CHART_AXIS_PRIMARY_Y, y1_min, y1_max);
+        if (chart_ecg_update == true)
+        {
+            lv_chart_set_range(chart_ecg, LV_CHART_AXIS_PRIMARY_Y, y_min_ecg, y_max_ecg);
+        }
 
         gx = 0;
-        y1_max = -100000;
-        y1_min = 100000;
 
-        /*y2_max = -100000;
-        y2_min = 100000;
-
-        y3_max = -100000;
-        y3_min = 100000;*/
+        y_max_ecg = -900000;
+        y_min_ecg = 900000;
     }
 }
 
-void scr_ecg_plot_ecg(float plot_data)
+void hpi_ecg_disp_draw_plotECG(int32_t *data_ecg, int num_samples, bool ecg_lead_off)
 {
-    if (chart1_update == true)
+    if (chart_ecg_update == true) // && ecg_lead_off == false)
     {
-
-        if (plot_data < y1_min)
+        for (int i = 0; i < num_samples; i++)
         {
-            y1_min = plot_data;
-        }
+            int32_t data_ecg_i = (data_ecg[i]/1000);
 
-        if (plot_data > y1_max)
-        {
-            y1_max = plot_data;
-        }
+            if (data_ecg_i < y_min_ecg)
+            {
+                y_min_ecg = data_ecg_i;
+            }
 
-        // printk("E");
-        lv_chart_set_next_value(chart1, ser1, plot_data);
-        hpi_ecg_add_samples(1);
-        hpi_ecg_do_set_scale();
+            if (data_ecg_i > y_max_ecg)
+            {
+                y_max_ecg = data_ecg_i;
+            }
+
+            /*if (ecg_plot_hidden == true)
+            {
+                lv_obj_add_flag(label_ecg_lead_off, LV_OBJ_FLAG_HIDDEN);
+                lv_obj_clear_flag(chart_ecg, LV_OBJ_FLAG_HIDDEN);
+                ecg_plot_hidden = false;
+            }*/
+
+            // printk("E");
+
+            lv_chart_set_next_value(chart_ecg, ser_ecg, data_ecg_i);
+            hpi_ecg_disp_add_samples(1);
+            hpi_ecg_disp_do_set_scale(DISP_WINDOW_SIZE_ECG);
+        }
+        // lv_chart_set_next_value(chart_ecg, ser_ecg, data_ecg);
+        // hpi_ecg_disp_add_samples(1);
+        // hpi_ecg_disp_do_set_scale(DISP_WINDOW_SIZE_ECG);
     }
+    /*else if (ecg_lead_off == true)
+    {
+        lv_obj_add_flag(chart_ecg, LV_OBJ_FLAG_HIDDEN);
+        lv_obj_clear_flag(label_ecg_lead_off, LV_OBJ_FLAG_HIDDEN);
+        ecg_plot_hidden = true;
+    }*/
 }
+
